Backport PR 1864 (#131)
* test: show issues with deduplication

* Refactor and fix deduplication in type enclosing.

Deduplication can only be done if the type have been printed. We perform that printing only at the junction between the reconstructed identifier enclosings and the ones from the tree nodes because we often want to keep both. All other duplicated ranges are removed.

* Large index alwways show the last type.

* Type-enclosing: revert deduplication in most cases

Since deduplciation has been introduced a while ago the results of type-enclosing (the number of enclosings) have been unstable. This cannot be solved easily on the server-side without printing more types which can lead to performance issues when large modules are involed. We now leave the responsability of deduplication to the clients.

* Promote tests after performing backport

* Stop deduplicating all results

---------

Co-authored-by: Ulysse Gérard <thevoodoos@gmail.com>